requirements.txt
pygame>=2.5
neat-python
